create a new package io.delta.sources, and merge the tests of the implicit sources in one test
