+++ src/main/scala/org/apache/spark/sql/delta/sources/DeltaDataSource.scala
@@ -331,56 +331,4 @@ object DeltaDataSource extends DatabricksLogging {
       None
     }
   }
-
-  /**
-   * Extends the DataFrameReader API by adding a delta function
-   * Usage:
-   * {{{
-   * spark.read.delta(path)
-   * }}}
-   */
-  implicit class DeltaDataFrameReader(val reader: DataFrameReader) extends AnyVal {
-    def delta(path: String): DataFrame = {
-      reader.format("delta").load(path)
-    }
-  }
-
-  /**
-   * Extends the DataFrameReader API by adding a delta function
-   * Usage: spark.readStream.delta(path)
-   * {{{
-   * spark.readStream.delta(path)
-   * }}}
-   */
-  implicit class DeltaStreamReader(val dataStreamReader: DataStreamReader) extends AnyVal {
-    def delta(path: String): DataFrame = {
-      dataStreamReader.format("delta").load(path)
-    }
-  }
-
-  /**
-   * Extends the DataFrameWriter API by adding a delta function
-   * Usage:
-   * {{{
-   * df.write.delta(path)
-   * }}}
-   */
-  implicit class DeltaWriter[T](val dfWriter: DataFrameWriter[T]) extends AnyVal {
-    def delta(output: String): Unit = {
-      dfWriter.format("delta").save(output)
-    }
-  }
-
-  /**
-   * Extends the DataStreamWriter API by adding a delta function
-   * Usage:
-   * {{{
-   * ds.writeStream.delta(path)
-   * }}}
-   */
-  implicit class DeltaStreamWriter[T](val dataStreamWriter: DataStreamWriter[T]) extends AnyVal {
-    def delta(path: String): StreamingQuery = {
-      dataStreamWriter.format("delta").start(path)
-    }
-  }
 }
